requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
email-validator==1.1.3
Flask-Mail==0.9.1
gunicorn==20.1.0
psycopg2-binary==2.9.1 